Ejemplo, algoritmo dinamico, filas columnas
