+++ secuencial/bin/Debug/netcoreapp3.1/Secuencial.deps.json
@@ -1,0 +1,23 @@
+{
+  "runtimeTarget": {
+    "name": ".NETCoreApp,Version=v3.1",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETCoreApp,Version=v3.1": {
+      "Secuencial/1.0.0": {
+        "runtime": {
+          "Secuencial.dll": {}
+        }
+      }
+    }
+  },
+  "libraries": {
+    "Secuencial/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
